Expect CTestConfig.cmake in either the build or source directory
Provide the ability to configure CTest with setting different from the one
available in the source tree by checking first if CTestConfig.cmake
exists in the build tree first.

The motivation is to allow build system checking out external project to
test and/or package them and submit the associated results to a different
dashboard than the one specified (or not) in the source of the external
project.

For example, the build system of Slicer can checkout, build, test
and package what I will  call "extension". These extensions can be developed
by third party who can test and submit to their own dashboard / project.
When checkout by Slicer build system, the default dashboard can be now
overwritten by adding a custom CTestConfig.cmake to the build directory.
And if not overwritten, it would avoid to create CTestConfig.cmake within
the source checkout of the extension.